feat: markup done in lis st pr by pr id and inventory sum fixed

+++ domain/store_product.model.go
@@ -18,7 +18,7 @@ type StoreProduct struct {
 	BonusAmount         float64    `gorm:"bonus_amount" json:"bonus_amount"`
 	BonusPercent        int        `gorm:"bonus_percent" json:"bonus_percent"`
 	Vat                 int        `gorm:"vat" json:"vat"`
-	Markup              int        `gorm:"markup" json:"markup"`
+	Markup              float64    `gorm:"markup" json:"markup"`
 	ShortName           string     `gorm:"short_name" json:"short_name"`
 	ExpireDate          *time.Time `gorm:"expire_date" json:"expire_date"`
 	CreatedAt           *time.Time `gorm:"created_at" json:"created_at"`
